feat(ios-google): move initial region logic to layoutSubviews
The logic for setting the initial region for the map was previously moved to didPrepareMap in an
attempt to fix the initialRegion prop. The mapView is yet to receive its frame in didMoveToWindow,
so calculating the zoom level fails. The mapView should have its frame in layoutSubviews,
which will be fired earlier than didPrepareMap, so moving it here makes sense.